Add IShaderCompiler api to bob for custom compilation of shaders
