+++ com.dynamo.cr/com.dynamo.cr.bob/.gitignore
@@ -20,6 +20,7 @@
 /libexec/*/*luajit*
 /libexec/*/glslc*
 /libexec/*/spirv-cross*
+/libexec/*/spirv-opt*
 /libexec/*/luac-32*
 /share/luajit
 /res
